Fold functionality into numba.testing
